pyproject.toml
[tool.poetry]
name = "stock-evaluator"
version = "0.1.0"
description = "Web Application for Stock Evaluation using Monte Carlo Simulations, Black-Scholes-Merton Model for Stock Options, and Sentiment Analysis."
authors = ["xavsant <xavierboons2021@economics.smu.edu.sg>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
numpy = "1.26.4"
streamlit = "^1.39.0"
fastapi = {extras = ["standard"], version = "^0.115.4"}
yfinance = "^0.2.48"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
selenium = "^4.26.1"
requests = "^2.32.3"
nltk = "^3.9.1"
contractions = "^0.1.73"
ipykernel = "^6.29.5"
scipy = "1.13.1"
matplotlib = "^3.9.2"
plotly = "^5.24.1"
nbformat = "^5.10.4"
seaborn = "^0.13.2"
gensim = "^4.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
